Removes the testing commands that were hacked into ServerBossRoomState
Moves the testing commands into a separate component (on the same prefab).
