+++ Assets/BossRoom/Scripts/Server/ServerTestingHotkeys.cs
@@ -1,0 +1,67 @@
+using System.Collections;
+using System.Collections.Generic;
+using MLAPI;
+using UnityEngine;
+
+namespace BossRoom.Server
+{
+    /// <summary>
+    /// Provides various special commands that the Host can use while developing and
+    /// debugging the game. Remove this component before releasing your game! (So that
+    /// regular players can't spawn extra bosses etc.)
+    /// </summary>
+    public class ServerTestingHotkeys : NetworkBehaviour
+    {
+        [SerializeField]
+        [Tooltip("Enemy to spawn. Make sure this is included in the NetworkManager's list of prefabs!")]
+        private NetworkObject m_EnemyPrefab;
+
+        [SerializeField]
+        [Tooltip("Boss to spawn. Make sure this is included in the NetworkManager's list of prefabs!")]
+        private NetworkObject m_BossPrefab;
+
+        [SerializeField]
+        [Tooltip("Key that the Host can press to spawn an extra enemy")]
+        KeyCode m_SpawnEnemyKeyCode = KeyCode.E;
+
+        [SerializeField]
+        [Tooltip("Key that the Host can press to spawn an extra boss")]
+        KeyCode m_SpawnBossKeyCode = KeyCode.B;
+
+        [SerializeField]
+        [Tooltip("Key that the Host can press to quit the game")]
+        KeyCode m_InstantQuitKeyCode = KeyCode.Q;
+
+        public override void NetworkStart()
+        {
+            base.NetworkStart();
+
+            if (!IsServer)
+            {
+                // these commands don't work on the client
+                enabled = false;
+            }
+        }
+
+        private void Update()
+        {
+            if (!IsServer) { return; } // not initialized yet
+
+            if (m_SpawnEnemyKeyCode != KeyCode.None && Input.GetKeyDown(m_SpawnEnemyKeyCode))
+            {
+                var newEnemy = Instantiate(m_EnemyPrefab);
+                newEnemy.SpawnWithOwnership(NetworkManager.Singleton.LocalClientId);
+            }
+            if (m_SpawnBossKeyCode != KeyCode.None && Input.GetKeyDown(m_SpawnBossKeyCode))
+            {
+                var newEnemy = Instantiate(m_BossPrefab);
+                newEnemy.SpawnWithOwnership(NetworkManager.Singleton.LocalClientId);
+            }
+            if (m_InstantQuitKeyCode != KeyCode.None && Input.GetKeyDown(m_InstantQuitKeyCode))
+            {
+                GameStateRelay.SetRelayObject(false); // indicate to the post-game screen that the game was lost
+                MLAPI.SceneManagement.NetworkSceneManager.SwitchScene("PostGame");
+            }
+        }
+    }
+}
