+++ Assets/BossRoom/Scripts/Server/ServerTestingHotkeys.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab1e76745edfc434ab8154ad27efc5fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
